requirements.txt
# FastAPI y servidor
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# RAG y embeddings - PostgreSQL con pgvector
openai>=1.3.0
pgvector==0.2.4
cloud-sql-python-connector[pg8000]==1.18.3
sqlalchemy>=2.0.30
pg8000>=1.29.8

# Procesamiento de datos básico
numpy>=1.24.3
pandas>=2.1.4
unidecode>=1.3.8

# Utilidades HTTP
requests>=2.31.0
aiohttp>=3.8.1
pytz>=2023.3
tenacity>=8.2.0

# Telegram Bot API
python-telegram-bot>=20.7

# Google APIs
google-api-python-client==2.118.0
google-cloud-storage>=2.13.0
google-cloud-logging>=3.5.0
google-cloud-secret-manager>=2.0.0 
PyYAML>=6.0.1